Form changes now trigger learning EVOLVE_MOVEs
