+++ src/data/balance/pokemon-level-moves.ts
@@ -11,7 +11,7 @@ interface PokemonFormLevelMoves {
   [key: number]: LevelMoves;
 }
 
-interface PokemonSpeciesFormLevelMoves {
+export interface PokemonSpeciesFormLevelMoves {
   [key: number]: PokemonFormLevelMoves;
 }
 
@@ -18925,1016 +18925,4 @@ export const pokemonSpeciesLevelMoves: PokemonSpeciesLevelMoves = {
     [64, MoveId.HAMMER_ARM],
     [70, MoveId.BLOOD_MOON],
   ],
-};
-
-export const pokemonFormLevelMoves: PokemonSpeciesFormLevelMoves = {
-  [Species.PIKACHU]: {
-    // Custom
-    1: [
-      [1, MoveId.TAIL_WHIP],
-      [1, MoveId.GROWL],
-      [1, MoveId.THUNDER_SHOCK],
-      [1, MoveId.QUICK_ATTACK],
-      [1, MoveId.SWEET_KISS],
-      [1, MoveId.CHARM],
-      [1, MoveId.NASTY_PLOT],
-      [1, MoveId.PLAY_NICE],
-      [1, MoveId.NUZZLE],
-      [4, MoveId.THUNDER_WAVE],
-      [8, MoveId.DOUBLE_TEAM],
-      [12, MoveId.ELECTRO_BALL],
-      [16, MoveId.FEINT],
-      [20, MoveId.ZIPPY_ZAP], // Custom
-      [24, MoveId.AGILITY],
-      [28, MoveId.IRON_TAIL],
-      [32, MoveId.DISCHARGE],
-      [34, MoveId.FLOATY_FALL], // Custom
-      [36, MoveId.THUNDERBOLT],
-      [40, MoveId.LIGHT_SCREEN],
-      [42, MoveId.SPLISHY_SPLASH], // Custom
-      [44, MoveId.THUNDER],
-      [48, MoveId.PIKA_PAPOW],
-    ],
-    2: [
-      [1, MoveId.TAIL_WHIP],
-      [1, MoveId.GROWL],
-      [1, MoveId.THUNDER_SHOCK],
-      [1, MoveId.QUICK_ATTACK],
-      [1, MoveId.SWEET_KISS],
-      [1, MoveId.CHARM],
-      [1, MoveId.NASTY_PLOT],
-      [1, MoveId.PLAY_NICE],
-      [1, MoveId.NUZZLE],
-      [4, MoveId.THUNDER_WAVE],
-      [8, MoveId.DOUBLE_TEAM],
-      [12, MoveId.ELECTRO_BALL],
-      [16, MoveId.FEINT],
-      [20, MoveId.SPARK],
-      [24, MoveId.AGILITY],
-      [28, MoveId.IRON_TAIL],
-      [32, MoveId.DISCHARGE],
-      [36, MoveId.THUNDERBOLT],
-      [40, MoveId.LIGHT_SCREEN],
-      [44, MoveId.THUNDER],
-      [48, MoveId.PIKA_PAPOW],
-    ],
-    3: [
-      [1, MoveId.METEOR_MASH],
-      [1, MoveId.TAIL_WHIP],
-      [1, MoveId.GROWL],
-      [1, MoveId.THUNDER_SHOCK],
-      [1, MoveId.QUICK_ATTACK],
-      [1, MoveId.SWEET_KISS],
-      [1, MoveId.CHARM],
-      [1, MoveId.NASTY_PLOT],
-      [1, MoveId.PLAY_NICE],
-      [1, MoveId.NUZZLE],
-      [4, MoveId.THUNDER_WAVE],
-      [8, MoveId.DOUBLE_TEAM],
-      [12, MoveId.ELECTRO_BALL],
-      [16, MoveId.FEINT],
-      [20, MoveId.SPARK],
-      [24, MoveId.AGILITY],
-      [28, MoveId.IRON_TAIL],
-      [32, MoveId.DISCHARGE],
-      [36, MoveId.THUNDERBOLT],
-      [40, MoveId.LIGHT_SCREEN],
-      [44, MoveId.THUNDER],
-      [48, MoveId.PIKA_PAPOW],
-    ],
-    4: [
-      [1, MoveId.ICICLE_CRASH],
-      [1, MoveId.TAIL_WHIP],
-      [1, MoveId.GROWL],
-      [1, MoveId.THUNDER_SHOCK],
-      [1, MoveId.QUICK_ATTACK],
-      [1, MoveId.SWEET_KISS],
-      [1, MoveId.CHARM],
-      [1, MoveId.NASTY_PLOT],
-      [1, MoveId.PLAY_NICE],
-      [1, MoveId.NUZZLE],
-      [4, MoveId.THUNDER_WAVE],
-      [8, MoveId.DOUBLE_TEAM],
-      [12, MoveId.ELECTRO_BALL],
-      [16, MoveId.FEINT],
-      [20, MoveId.SPARK],
-      [24, MoveId.AGILITY],
-      [28, MoveId.IRON_TAIL],
-      [32, MoveId.DISCHARGE],
-      [36, MoveId.THUNDERBOLT],
-      [40, MoveId.LIGHT_SCREEN],
-      [44, MoveId.THUNDER],
-      [48, MoveId.PIKA_PAPOW],
-    ],
-    5: [
-      [1, MoveId.DRAINING_KISS],
-      [1, MoveId.TAIL_WHIP],
-      [1, MoveId.GROWL],
-      [1, MoveId.THUNDER_SHOCK],
-      [1, MoveId.QUICK_ATTACK],
-      [1, MoveId.SWEET_KISS],
-      [1, MoveId.CHARM],
-      [1, MoveId.NASTY_PLOT],
-      [1, MoveId.PLAY_NICE],
-      [1, MoveId.NUZZLE],
-      [4, MoveId.THUNDER_WAVE],
-      [8, MoveId.DOUBLE_TEAM],
-      [12, MoveId.ELECTRO_BALL],
-      [16, MoveId.FEINT],
-      [20, MoveId.SPARK],
-      [24, MoveId.AGILITY],
-      [28, MoveId.IRON_TAIL],
-      [32, MoveId.DISCHARGE],
-      [36, MoveId.THUNDERBOLT],
-      [40, MoveId.LIGHT_SCREEN],
-      [44, MoveId.THUNDER],
-      [48, MoveId.PIKA_PAPOW],
-    ],
-    6: [
-      [1, MoveId.ELECTRIC_TERRAIN],
-      [1, MoveId.TAIL_WHIP],
-      [1, MoveId.GROWL],
-      [1, MoveId.THUNDER_SHOCK],
-      [1, MoveId.QUICK_ATTACK],
-      [1, MoveId.SWEET_KISS],
-      [1, MoveId.CHARM],
-      [1, MoveId.NASTY_PLOT],
-      [1, MoveId.PLAY_NICE],
-      [1, MoveId.NUZZLE],
-      [4, MoveId.THUNDER_WAVE],
-      [8, MoveId.DOUBLE_TEAM],
-      [12, MoveId.ELECTRO_BALL],
-      [16, MoveId.FEINT],
-      [20, MoveId.SPARK],
-      [24, MoveId.AGILITY],
-      [28, MoveId.IRON_TAIL],
-      [32, MoveId.DISCHARGE],
-      [36, MoveId.THUNDERBOLT],
-      [40, MoveId.LIGHT_SCREEN],
-      [44, MoveId.THUNDER],
-      [48, MoveId.PIKA_PAPOW],
-    ],
-    7: [
-      [1, MoveId.FLYING_PRESS],
-      [1, MoveId.TAIL_WHIP],
-      [1, MoveId.GROWL],
-      [1, MoveId.THUNDER_SHOCK],
-      [1, MoveId.QUICK_ATTACK],
-      [1, MoveId.SWEET_KISS],
-      [1, MoveId.CHARM],
-      [1, MoveId.NASTY_PLOT],
-      [1, MoveId.PLAY_NICE],
-      [1, MoveId.NUZZLE],
-      [4, MoveId.THUNDER_WAVE],
-      [8, MoveId.DOUBLE_TEAM],
-      [12, MoveId.ELECTRO_BALL],
-      [16, MoveId.FEINT],
-      [20, MoveId.SPARK],
-      [24, MoveId.AGILITY],
-      [28, MoveId.IRON_TAIL],
-      [32, MoveId.DISCHARGE],
-      [36, MoveId.THUNDERBOLT],
-      [40, MoveId.LIGHT_SCREEN],
-      [44, MoveId.THUNDER],
-      [48, MoveId.PIKA_PAPOW],
-    ],
-    8: [
-      [1, MoveId.TAIL_WHIP],
-      [1, MoveId.GROWL],
-      [1, MoveId.THUNDER_SHOCK],
-      [1, MoveId.QUICK_ATTACK],
-      [1, MoveId.SWEET_KISS],
-      [1, MoveId.CHARM],
-      [1, MoveId.NASTY_PLOT],
-      [1, MoveId.PLAY_NICE],
-      [1, MoveId.NUZZLE],
-      [4, MoveId.THUNDER_WAVE],
-      [8, MoveId.DOUBLE_TEAM],
-      [12, MoveId.ELECTRO_BALL],
-      [16, MoveId.FEINT],
-      [20, MoveId.ZIPPY_ZAP], // Custom
-      [24, MoveId.AGILITY],
-      [28, MoveId.IRON_TAIL],
-      [32, MoveId.DISCHARGE],
-      [34, MoveId.FLOATY_FALL], // Custom
-      [36, MoveId.THUNDERBOLT],
-      [40, MoveId.LIGHT_SCREEN],
-      [42, MoveId.SPLISHY_SPLASH], // Custom
-      [44, MoveId.THUNDER],
-      [48, MoveId.PIKA_PAPOW],
-    ],
-  },
-  [Species.EEVEE]: {
-    // Custom
-    1: [
-      [1, MoveId.TACKLE],
-      [1, MoveId.TAIL_WHIP],
-      [1, MoveId.GROWL],
-      [1, MoveId.HELPING_HAND],
-      [1, MoveId.COVET],
-      [5, MoveId.SAND_ATTACK],
-      [10, MoveId.QUICK_ATTACK],
-      [15, MoveId.BABY_DOLL_EYES],
-      [18, MoveId.BOUNCY_BUBBLE], // Custom
-      [18, MoveId.SIZZLY_SLIDE], // Custom
-      [18, MoveId.BUZZY_BUZZ], // Custom
-      [20, MoveId.SWIFT],
-      [25, MoveId.BITE],
-      [30, MoveId.COPYCAT],
-      [33, MoveId.BADDY_BAD], // Custom
-      [33, MoveId.GLITZY_GLOW], // Custom
-      [35, MoveId.BATON_PASS],
-      [40, MoveId.VEEVEE_VOLLEY], // Custom, replaces Take Down
-      [43, MoveId.FREEZY_FROST], // Custom
-      [43, MoveId.SAPPY_SEED], // Custom
-      [45, MoveId.CHARM],
-      [50, MoveId.DOUBLE_EDGE],
-      [53, MoveId.SPARKLY_SWIRL], // Custom
-      [55, MoveId.LAST_RESORT],
-    ],
-    2: [
-      [1, MoveId.TACKLE],
-      [1, MoveId.TAIL_WHIP],
-      [1, MoveId.GROWL],
-      [1, MoveId.HELPING_HAND],
-      [1, MoveId.COVET],
-      [5, MoveId.SAND_ATTACK],
-      [10, MoveId.QUICK_ATTACK],
-      [15, MoveId.BABY_DOLL_EYES],
-      [18, MoveId.BOUNCY_BUBBLE], // Custom
-      [18, MoveId.SIZZLY_SLIDE], // Custom
-      [18, MoveId.BUZZY_BUZZ], // Custom
-      [20, MoveId.SWIFT],
-      [25, MoveId.BITE],
-      [30, MoveId.COPYCAT],
-      [33, MoveId.BADDY_BAD], // Custom
-      [33, MoveId.GLITZY_GLOW], // Custom
-      [35, MoveId.BATON_PASS],
-      [40, MoveId.VEEVEE_VOLLEY], // Custom, replaces Take Down
-      [43, MoveId.FREEZY_FROST], // Custom
-      [43, MoveId.SAPPY_SEED], // Custom
-      [45, MoveId.CHARM],
-      [50, MoveId.DOUBLE_EDGE],
-      [53, MoveId.SPARKLY_SWIRL], // Custom
-      [55, MoveId.LAST_RESORT],
-    ],
-  },
-  [Species.DEOXYS]: {
-    1: [
-      [1, MoveId.CONFUSION], // Custom
-      [1, MoveId.WRAP],
-      [1, MoveId.LEER],
-      [7, MoveId.NIGHT_SHADE],
-      [13, MoveId.TELEPORT],
-      [19, MoveId.TAUNT],
-      [25, MoveId.PURSUIT],
-      [31, MoveId.PSYCHIC],
-      [37, MoveId.SUPERPOWER],
-      [43, MoveId.PSYCHO_SHIFT],
-      [49, MoveId.ZEN_HEADBUTT],
-      [55, MoveId.COSMIC_POWER],
-      [61, MoveId.ZAP_CANNON],
-      [67, MoveId.PSYCHO_BOOST],
-      [73, MoveId.HYPER_BEAM],
-    ],
-    2: [
-      [1, MoveId.CONFUSION], // Custom
-      [1, MoveId.WRAP],
-      [1, MoveId.LEER],
-      [7, MoveId.NIGHT_SHADE],
-      [13, MoveId.TELEPORT],
-      [19, MoveId.KNOCK_OFF],
-      [25, MoveId.SPIKES],
-      [31, MoveId.PSYCHIC],
-      [37, MoveId.SNATCH],
-      [43, MoveId.PSYCHO_SHIFT],
-      [49, MoveId.ZEN_HEADBUTT],
-      [55, MoveId.AMNESIA],
-      [55, MoveId.IRON_DEFENSE],
-      [61, MoveId.RECOVER],
-      [67, MoveId.PSYCHO_BOOST],
-      [73, MoveId.COUNTER],
-      [73, MoveId.MIRROR_COAT],
-    ],
-    3: [
-      [1, MoveId.CONFUSION], // Custom
-      [1, MoveId.WRAP],
-      [1, MoveId.LEER],
-      [7, MoveId.NIGHT_SHADE],
-      [13, MoveId.DOUBLE_TEAM],
-      [19, MoveId.KNOCK_OFF],
-      [25, MoveId.PURSUIT],
-      [31, MoveId.PSYCHIC],
-      [37, MoveId.SWIFT],
-      [43, MoveId.PSYCHO_SHIFT],
-      [49, MoveId.ZEN_HEADBUTT],
-      [55, MoveId.AGILITY],
-      [61, MoveId.RECOVER],
-      [67, MoveId.PSYCHO_BOOST],
-      [73, MoveId.EXTREME_SPEED],
-    ],
-  },
-  [Species.WORMADAM]: {
-    1: [
-      [EVOLVE_MOVE, MoveId.QUIVER_DANCE],
-      [1, MoveId.STRUGGLE_BUG], // Previous Stage Move, Custom
-      [1, MoveId.TACKLE],
-      [1, MoveId.PROTECT],
-      [1, MoveId.SUCKER_PUNCH],
-      [1, MoveId.BUG_BITE],
-      [1, MoveId.PROTECT],
-      [10, MoveId.TACKLE],
-      [20, MoveId.STRING_SHOT],
-      [23, MoveId.CONFUSION],
-      [26, MoveId.ROCK_BLAST],
-      [29, MoveId.HARDEN],
-      [32, MoveId.PSYBEAM],
-      [35, MoveId.INFESTATION],
-      [38, MoveId.FLAIL],
-      [41, MoveId.ATTRACT],
-      [44, MoveId.PSYCHIC],
-      [47, MoveId.FISSURE],
-      [50, MoveId.BUG_BUZZ],
-    ],
-    2: [
-      [EVOLVE_MOVE, MoveId.QUIVER_DANCE],
-      [1, MoveId.STRUGGLE_BUG], // Previous Stage Move, Custom
-      [1, MoveId.METAL_BURST],
-      [1, MoveId.TACKLE],
-      [1, MoveId.PROTECT],
-      [1, MoveId.SUCKER_PUNCH],
-      [1, MoveId.BUG_BITE],
-      [1, MoveId.PROTECT],
-      [10, MoveId.TACKLE],
-      [20, MoveId.STRING_SHOT],
-      [23, MoveId.CONFUSION],
-      [26, MoveId.METAL_BURST],
-      [29, MoveId.METAL_SOUND],
-      [32, MoveId.PSYBEAM],
-      [35, MoveId.INFESTATION],
-      [38, MoveId.FLAIL],
-      [41, MoveId.ATTRACT],
-      [44, MoveId.PSYCHIC],
-      [47, MoveId.IRON_HEAD],
-      [50, MoveId.BUG_BUZZ],
-    ],
-  },
-  [Species.ROTOM]: {
-    1: [
-      [1, MoveId.OVERHEAT],
-      [1, MoveId.DOUBLE_TEAM],
-      [1, MoveId.ASTONISH],
-      [5, MoveId.THUNDER_SHOCK],
-      [10, MoveId.CONFUSE_RAY],
-      [15, MoveId.CHARGE],
-      [20, MoveId.ELECTRO_BALL],
-      [25, MoveId.THUNDER_WAVE],
-      [30, MoveId.SHOCK_WAVE],
-      [35, MoveId.HEX],
-      [40, MoveId.SUBSTITUTE],
-      [45, MoveId.TRICK],
-      [50, MoveId.DISCHARGE],
-      [55, MoveId.UPROAR],
-    ],
-    2: [
-      [1, MoveId.HYDRO_PUMP],
-      [1, MoveId.DOUBLE_TEAM],
-      [1, MoveId.ASTONISH],
-      [5, MoveId.THUNDER_SHOCK],
-      [10, MoveId.CONFUSE_RAY],
-      [15, MoveId.CHARGE],
-      [20, MoveId.ELECTRO_BALL],
-      [25, MoveId.THUNDER_WAVE],
-      [30, MoveId.SHOCK_WAVE],
-      [35, MoveId.HEX],
-      [40, MoveId.SUBSTITUTE],
-      [45, MoveId.TRICK],
-      [50, MoveId.DISCHARGE],
-      [55, MoveId.UPROAR],
-    ],
-    3: [
-      [1, MoveId.BLIZZARD],
-      [1, MoveId.DOUBLE_TEAM],
-      [1, MoveId.ASTONISH],
-      [5, MoveId.THUNDER_SHOCK],
-      [10, MoveId.CONFUSE_RAY],
-      [15, MoveId.CHARGE],
-      [20, MoveId.ELECTRO_BALL],
-      [25, MoveId.THUNDER_WAVE],
-      [30, MoveId.SHOCK_WAVE],
-      [35, MoveId.HEX],
-      [40, MoveId.SUBSTITUTE],
-      [45, MoveId.TRICK],
-      [50, MoveId.DISCHARGE],
-      [55, MoveId.UPROAR],
-    ],
-    4: [
-      [1, MoveId.AIR_SLASH],
-      [1, MoveId.DOUBLE_TEAM],
-      [1, MoveId.ASTONISH],
-      [5, MoveId.THUNDER_SHOCK],
-      [10, MoveId.CONFUSE_RAY],
-      [15, MoveId.CHARGE],
-      [20, MoveId.ELECTRO_BALL],
-      [25, MoveId.THUNDER_WAVE],
-      [30, MoveId.SHOCK_WAVE],
-      [35, MoveId.HEX],
-      [40, MoveId.SUBSTITUTE],
-      [45, MoveId.TRICK],
-      [50, MoveId.DISCHARGE],
-      [55, MoveId.UPROAR],
-    ],
-    5: [
-      [1, MoveId.LEAF_STORM],
-      [1, MoveId.DOUBLE_TEAM],
-      [1, MoveId.ASTONISH],
-      [5, MoveId.THUNDER_SHOCK],
-      [10, MoveId.CONFUSE_RAY],
-      [15, MoveId.CHARGE],
-      [20, MoveId.ELECTRO_BALL],
-      [25, MoveId.THUNDER_WAVE],
-      [30, MoveId.SHOCK_WAVE],
-      [35, MoveId.HEX],
-      [40, MoveId.SUBSTITUTE],
-      [45, MoveId.TRICK],
-      [50, MoveId.DISCHARGE],
-      [55, MoveId.UPROAR],
-    ],
-  },
-  [Species.SHAYMIN]: {
-    1: [
-      [1, MoveId.LEAFAGE], // Custom
-      [1, MoveId.GROWTH],
-      [10, MoveId.MAGICAL_LEAF],
-      [19, MoveId.LEECH_SEED],
-      [28, MoveId.QUICK_ATTACK],
-      [37, MoveId.SWEET_SCENT],
-      [46, MoveId.NATURAL_GIFT],
-      [55, MoveId.WORRY_SEED],
-      [64, MoveId.AIR_SLASH],
-      [73, MoveId.ENERGY_BALL],
-      [82, MoveId.SWEET_KISS],
-      [91, MoveId.LEAF_STORM],
-      [100, MoveId.SEED_FLARE],
-    ],
-  },
-  [Species.KYUREM]: {
-    1: [
-      [1, MoveId.DRAGON_BREATH],
-      [1, MoveId.ANCIENT_POWER],
-      [1, MoveId.NOBLE_ROAR],
-      [1, MoveId.FREEZE_DRY],
-      [8, MoveId.SLASH],
-      [16, MoveId.ENDEAVOR],
-      [24, MoveId.DRAGON_PULSE],
-      [32, MoveId.ICE_BEAM],
-      [40, MoveId.HYPER_VOICE],
-      [48, MoveId.FUSION_BOLT],
-      [56, MoveId.BLIZZARD],
-      [64, MoveId.IMPRISON],
-      [72, MoveId.OUTRAGE],
-      [80, MoveId.FREEZE_SHOCK],
-      [88, MoveId.SHEER_COLD],
-    ],
-    2: [
-      [1, MoveId.DRAGON_BREATH],
-      [1, MoveId.ANCIENT_POWER],
-      [1, MoveId.NOBLE_ROAR],
-      [1, MoveId.FREEZE_DRY],
-      [8, MoveId.SLASH],
-      [16, MoveId.ENDEAVOR],
-      [24, MoveId.DRAGON_PULSE],
-      [32, MoveId.ICE_BEAM],
-      [40, MoveId.HYPER_VOICE],
-      [48, MoveId.FUSION_FLARE],
-      [56, MoveId.BLIZZARD],
-      [64, MoveId.IMPRISON],
-      [72, MoveId.OUTRAGE],
-      [80, MoveId.ICE_BURN],
-      [88, MoveId.SHEER_COLD],
-    ],
-  },
-  [Species.MEOWSTIC]: {
-    1: [
-      [1, MoveId.SCRATCH],
-      [1, MoveId.LEER],
-      [1, MoveId.FAKE_OUT],
-      [1, MoveId.MAGICAL_LEAF],
-      [1, MoveId.DISARMING_VOICE],
-      [9, MoveId.CONFUSION],
-      [12, MoveId.STORED_POWER],
-      [15, MoveId.CHARGE_BEAM],
-      [18, MoveId.COVET],
-      [21, MoveId.PSYBEAM],
-      [24, MoveId.SUCKER_PUNCH],
-      [29, MoveId.ROLE_PLAY],
-      [34, MoveId.LIGHT_SCREEN],
-      [34, MoveId.REFLECT],
-      [39, MoveId.PSYSHOCK],
-      [44, MoveId.EXTRASENSORY],
-      [49, MoveId.SHADOW_BALL],
-      [54, MoveId.PSYCHIC],
-      [59, MoveId.FUTURE_SIGHT],
-    ],
-  },
-  [Species.HOOPA]: {
-    1: [
-      [1, MoveId.CONFUSION],
-      [1, MoveId.DESTINY_BOND],
-      [1, MoveId.ALLY_SWITCH],
-      [6, MoveId.ASTONISH],
-      [10, MoveId.TRICK],
-      [15, MoveId.LIGHT_SCREEN],
-      [19, MoveId.PSYBEAM],
-      [25, MoveId.SKILL_SWAP],
-      [29, MoveId.GUARD_SPLIT],
-      [29, MoveId.POWER_SPLIT],
-      [46, MoveId.KNOCK_OFF],
-      [50, MoveId.TRICK_ROOM],
-      [50, MoveId.WONDER_ROOM],
-      [55, MoveId.DARK_PULSE],
-      [75, MoveId.PSYCHIC],
-      [85, MoveId.HYPERSPACE_FURY],
-    ],
-  },
-  [Species.GRENINJA]: {
-    1: [
-      [EVOLVE_MOVE, MoveId.WATER_SHURIKEN],
-      [EVOLVE_MOVE, MoveId.WATER_SHURIKEN],
-      [1, MoveId.POUND],
-      [1, MoveId.GROWL],
-      [1, MoveId.WATER_GUN],
-      [1, MoveId.QUICK_ATTACK],
-      [1, MoveId.ROUND], // Previous Stage Move
-      [1, MoveId.FLING], // Previous Stage Move
-      [1, MoveId.SMACK_DOWN], // Previous Stage Move
-      [1, MoveId.BOUNCE], // Previous Stage Move
-      [1, MoveId.HAZE],
-      [1, MoveId.ROLE_PLAY],
-      [1, MoveId.NIGHT_SLASH],
-      [10, MoveId.LICK],
-      [14, MoveId.WATER_PULSE],
-      [19, MoveId.SMOKESCREEN],
-      [23, MoveId.SHADOW_SNEAK],
-      [28, MoveId.SPIKES],
-      [33, MoveId.AERIAL_ACE],
-      [42, MoveId.SUBSTITUTE],
-      [49, MoveId.EXTRASENSORY],
-      [56, MoveId.DOUBLE_TEAM],
-      [68, MoveId.HYDRO_PUMP],
-    ],
-  },
-  [Species.LYCANROC]: {
-    1: [
-      [EVOLVE_MOVE, MoveId.COUNTER],
-      [1, MoveId.SAND_ATTACK],
-      [1, MoveId.TACKLE],
-      [1, MoveId.LEER],
-      [1, MoveId.DOUBLE_TEAM],
-      [1, MoveId.REVERSAL],
-      [1, MoveId.ENDURE],
-      [1, MoveId.TAUNT],
-      [12, MoveId.ROCK_THROW],
-      [16, MoveId.HOWL],
-      [20, MoveId.BITE],
-      [24, MoveId.ROCK_TOMB],
-      [30, MoveId.ROAR],
-      [36, MoveId.ROCK_SLIDE],
-      [42, MoveId.CRUNCH],
-      [48, MoveId.SCARY_FACE],
-      [54, MoveId.STEALTH_ROCK],
-      [60, MoveId.STONE_EDGE],
-    ],
-    2: [
-      [EVOLVE_MOVE, MoveId.CRUSH_CLAW],
-      [1, MoveId.SAND_ATTACK],
-      [1, MoveId.TACKLE],
-      [1, MoveId.LEER],
-      [1, MoveId.COUNTER],
-      [1, MoveId.QUICK_ATTACK],
-      [1, MoveId.DOUBLE_TEAM],
-      [1, MoveId.REVERSAL],
-      [1, MoveId.ENDURE],
-      [1, MoveId.TAUNT],
-      [1, MoveId.SUCKER_PUNCH],
-      [1, MoveId.QUICK_GUARD],
-      [1, MoveId.ACCELEROCK],
-      [12, MoveId.ROCK_THROW],
-      [16, MoveId.HOWL],
-      [20, MoveId.BITE],
-      [24, MoveId.ROCK_TOMB],
-      [30, MoveId.ROAR],
-      [36, MoveId.ROCK_SLIDE],
-      [42, MoveId.CRUNCH],
-      [48, MoveId.SCARY_FACE],
-      [54, MoveId.STEALTH_ROCK],
-      [60, MoveId.STONE_EDGE],
-    ],
-  },
-  [Species.NECROZMA]: {
-    1: [
-      [EVOLVE_MOVE, MoveId.SUNSTEEL_STRIKE],
-      [1, MoveId.CONFUSION],
-      [1, MoveId.METAL_CLAW],
-      [1, MoveId.MORNING_SUN],
-      [1, MoveId.MOONLIGHT],
-      [1, MoveId.GRAVITY],
-      [1, MoveId.CHARGE_BEAM],
-      [8, MoveId.STEALTH_ROCK],
-      [16, MoveId.SLASH],
-      [24, MoveId.NIGHT_SLASH],
-      [32, MoveId.PSYCHO_CUT],
-      [40, MoveId.STORED_POWER],
-      [48, MoveId.ROCK_BLAST],
-      [56, MoveId.IRON_DEFENSE],
-      [64, MoveId.POWER_GEM],
-      [72, MoveId.PHOTON_GEYSER],
-      [80, MoveId.AUTOTOMIZE],
-      [88, MoveId.PRISMATIC_LASER],
-    ],
-    2: [
-      [EVOLVE_MOVE, MoveId.MOONGEIST_BEAM],
-      [1, MoveId.CONFUSION],
-      [1, MoveId.METAL_CLAW],
-      [1, MoveId.MORNING_SUN],
-      [1, MoveId.MOONLIGHT],
-      [1, MoveId.GRAVITY],
-      [1, MoveId.CHARGE_BEAM],
-      [8, MoveId.STEALTH_ROCK],
-      [16, MoveId.SLASH],
-      [24, MoveId.NIGHT_SLASH],
-      [32, MoveId.PSYCHO_CUT],
-      [40, MoveId.STORED_POWER],
-      [48, MoveId.ROCK_BLAST],
-      [56, MoveId.IRON_DEFENSE],
-      [64, MoveId.POWER_GEM],
-      [72, MoveId.PHOTON_GEYSER],
-      [80, MoveId.AUTOTOMIZE],
-      [88, MoveId.PRISMATIC_LASER],
-    ],
-    3: [
-      [EVOLVE_MOVE, MoveId.SUNSTEEL_STRIKE],
-      [EVOLVE_MOVE, MoveId.MOONGEIST_BEAM],
-      [1, MoveId.CONFUSION],
-      [1, MoveId.METAL_CLAW],
-      [1, MoveId.MORNING_SUN],
-      [1, MoveId.MOONLIGHT],
-      [1, MoveId.GRAVITY],
-      [1, MoveId.CHARGE_BEAM],
-      [8, MoveId.STEALTH_ROCK],
-      [16, MoveId.SLASH],
-      [24, MoveId.NIGHT_SLASH],
-      [32, MoveId.PSYCHO_CUT],
-      [40, MoveId.STORED_POWER],
-      [48, MoveId.ROCK_BLAST],
-      [56, MoveId.IRON_DEFENSE],
-      [64, MoveId.POWER_GEM],
-      [72, MoveId.PHOTON_GEYSER],
-      [80, MoveId.AUTOTOMIZE],
-      [88, MoveId.PRISMATIC_LASER],
-    ],
-  },
-  [Species.TOXTRICITY]: {
-    1: [
-      [EVOLVE_MOVE, MoveId.SPARK],
-      [1, MoveId.LEER],
-      [1, MoveId.GROWL],
-      [1, MoveId.ACID],
-      [1, MoveId.THUNDER_SHOCK],
-      [1, MoveId.FLAIL],
-      [1, MoveId.ACID_SPRAY],
-      [1, MoveId.BELCH],
-      [1, MoveId.NOBLE_ROAR],
-      [1, MoveId.NUZZLE],
-      [1, MoveId.TEARFUL_LOOK],
-      [4, MoveId.CHARGE],
-      [8, MoveId.SHOCK_WAVE],
-      [12, MoveId.SCARY_FACE],
-      [16, MoveId.TAUNT],
-      [24, MoveId.SCREECH],
-      [28, MoveId.SWAGGER],
-      [32, MoveId.TOXIC],
-      [36, MoveId.DISCHARGE],
-      [40, MoveId.POISON_JAB],
-      [44, MoveId.OVERDRIVE],
-      [48, MoveId.BOOMBURST],
-      [52, MoveId.MAGNETIC_FLUX],
-    ],
-  },
-  [Species.INDEEDEE]: {
-    1: [
-      [1, MoveId.STORED_POWER],
-      [1, MoveId.PLAY_NICE],
-      [5, MoveId.BATON_PASS],
-      [10, MoveId.DISARMING_VOICE],
-      [15, MoveId.PSYBEAM],
-      [20, MoveId.HELPING_HAND],
-      [25, MoveId.FOLLOW_ME],
-      [30, MoveId.HEALING_WISH],
-      [35, MoveId.PSYCHIC],
-      [40, MoveId.CALM_MIND],
-      [45, MoveId.GUARD_SPLIT],
-      [50, MoveId.PSYCHIC_TERRAIN],
-    ],
-  },
-  [Species.ZACIAN]: {
-    1: [
-      [EVOLVE_MOVE, MoveId.BEHEMOTH_BLADE],
-      [1, MoveId.BITE],
-      [1, MoveId.QUICK_ATTACK],
-      [1, MoveId.METAL_CLAW],
-      [1, MoveId.HOWL],
-      [1, MoveId.QUICK_GUARD],
-      [1, MoveId.SACRED_SWORD],
-      [11, MoveId.SLASH],
-      [22, MoveId.SWORDS_DANCE],
-      [33, MoveId.IRON_HEAD],
-      [44, MoveId.NOBLE_ROAR],
-      [55, MoveId.CRUNCH],
-      [66, MoveId.MOONBLAST],
-      [77, MoveId.CLOSE_COMBAT],
-      [88, MoveId.GIGA_IMPACT],
-    ],
-  },
-  [Species.ZAMAZENTA]: {
-    1: [
-      [EVOLVE_MOVE, MoveId.BEHEMOTH_BASH],
-      [1, MoveId.BITE],
-      [1, MoveId.QUICK_ATTACK],
-      [1, MoveId.METAL_CLAW],
-      [1, MoveId.HOWL],
-      [1, MoveId.WIDE_GUARD],
-      [11, MoveId.SLASH],
-      [22, MoveId.IRON_DEFENSE],
-      [33, MoveId.IRON_HEAD],
-      [44, MoveId.METAL_BURST],
-      [55, MoveId.CRUNCH],
-      [66, MoveId.MOONBLAST],
-      [77, MoveId.CLOSE_COMBAT],
-      [88, MoveId.GIGA_IMPACT],
-    ],
-  },
-  [Species.ETERNATUS]: {
-    1: [
-      [1, MoveId.AGILITY],
-      [1, MoveId.CONFUSE_RAY],
-      [1, MoveId.POISON_TAIL],
-      [1, MoveId.DRAGON_TAIL],
-      [8, MoveId.TOXIC],
-      [16, MoveId.VENOSHOCK],
-      [24, MoveId.DRAGON_DANCE],
-      [32, MoveId.CROSS_POISON],
-      [40, MoveId.DRAGON_PULSE],
-      [48, MoveId.FLAMETHROWER],
-      [56, MoveId.DYNAMAX_CANNON],
-      [64, MoveId.COSMIC_POWER],
-      [72, MoveId.RECOVER],
-      [80, MoveId.HYPER_BEAM],
-      [88, MoveId.ETERNABEAM],
-    ],
-  },
-  [Species.URSHIFU]: {
-    1: [
-      [EVOLVE_MOVE, MoveId.SURGING_STRIKES],
-      [1, MoveId.LEER],
-      [1, MoveId.FOCUS_ENERGY],
-      [1, MoveId.ENDURE],
-      [1, MoveId.ROCK_SMASH],
-      [1, MoveId.AQUA_JET],
-      [12, MoveId.AERIAL_ACE],
-      [16, MoveId.SCARY_FACE],
-      [20, MoveId.HEADBUTT],
-      [24, MoveId.BRICK_BREAK],
-      [28, MoveId.DETECT],
-      [32, MoveId.BULK_UP],
-      [36, MoveId.IRON_HEAD],
-      [40, MoveId.DYNAMIC_PUNCH],
-      [44, MoveId.COUNTER],
-      [48, MoveId.CLOSE_COMBAT],
-      [52, MoveId.FOCUS_PUNCH],
-    ],
-  },
-  [Species.CALYREX]: {
-    1: [
-      [1, MoveId.POUND],
-      [1, MoveId.SWORDS_DANCE],
-      [1, MoveId.STOMP],
-      [1, MoveId.DOUBLE_KICK],
-      [1, MoveId.TACKLE],
-      [1, MoveId.TAKE_DOWN],
-      [1, MoveId.THRASH],
-      [1, MoveId.DOUBLE_EDGE],
-      [1, MoveId.TAIL_WHIP],
-      [1, MoveId.MIST],
-      [1, MoveId.MEGA_DRAIN],
-      [1, MoveId.GROWTH],
-      [1, MoveId.CONFUSION],
-      [1, MoveId.TORMENT],
-      [1, MoveId.TAUNT],
-      [1, MoveId.IRON_DEFENSE],
-      [1, MoveId.AVALANCHE],
-      [1, MoveId.ICICLE_CRASH],
-      [1, MoveId.GLACIAL_LANCE],
-      [8, MoveId.LIFE_DEW],
-      [16, MoveId.GIGA_DRAIN],
-      [24, MoveId.PSYSHOCK],
-      [32, MoveId.HELPING_HAND],
-      [40, MoveId.AROMATHERAPY],
-      [40, MoveId.GRASSY_TERRAIN],
-      [48, MoveId.ENERGY_BALL],
-      [56, MoveId.PSYCHIC],
-      [64, MoveId.LEECH_SEED],
-      [72, MoveId.HEAL_PULSE],
-      [80, MoveId.SOLAR_BEAM],
-      [88, MoveId.FUTURE_SIGHT],
-    ],
-    2: [
-      [1, MoveId.POUND],
-      [1, MoveId.STOMP],
-      [1, MoveId.DOUBLE_KICK],
-      [1, MoveId.TACKLE],
-      [1, MoveId.TAKE_DOWN],
-      [1, MoveId.THRASH],
-      [1, MoveId.DOUBLE_EDGE],
-      [1, MoveId.TAIL_WHIP],
-      [1, MoveId.DISABLE],
-      [1, MoveId.MEGA_DRAIN],
-      [1, MoveId.GROWTH],
-      [1, MoveId.CONFUSION],
-      [1, MoveId.AGILITY],
-      [1, MoveId.CONFUSE_RAY],
-      [1, MoveId.HAZE],
-      [1, MoveId.SHADOW_BALL],
-      [1, MoveId.NASTY_PLOT],
-      [1, MoveId.HEX],
-      [1, MoveId.ASTRAL_BARRAGE],
-      [8, MoveId.LIFE_DEW],
-      [16, MoveId.GIGA_DRAIN],
-      [24, MoveId.PSYSHOCK],
-      [32, MoveId.HELPING_HAND],
-      [40, MoveId.GRASSY_TERRAIN],
-      [40, MoveId.PSYCHIC_TERRAIN],
-      [48, MoveId.ENERGY_BALL],
-      [56, MoveId.PSYCHIC],
-      [64, MoveId.LEECH_SEED],
-      [72, MoveId.HEAL_PULSE],
-      [80, MoveId.SOLAR_BEAM],
-      [88, MoveId.FUTURE_SIGHT],
-    ],
-  },
-  [Species.OINKOLOGNE]: {
-    1: [
-      [1, MoveId.TACKLE],
-      [1, MoveId.TAIL_WHIP],
-      [3, MoveId.DISARMING_VOICE],
-      [6, MoveId.ECHOED_VOICE],
-      [9, MoveId.MUD_SHOT],
-      [12, MoveId.COVET],
-      [15, MoveId.DIG],
-      [17, MoveId.HEADBUTT],
-      [23, MoveId.YAWN],
-      [28, MoveId.TAKE_DOWN],
-      [30, MoveId.WORK_UP],
-      [34, MoveId.UPROAR],
-      [39, MoveId.DOUBLE_EDGE],
-      [45, MoveId.EARTH_POWER],
-      [51, MoveId.BELCH],
-    ],
-  },
-  [Species.REVAVROOM]: {
-    1: [
-      [EVOLVE_MOVE, MoveId.WICKED_TORQUE],
-      [EVOLVE_MOVE, MoveId.SHIFT_GEAR],
-      [1, MoveId.LICK],
-      [1, MoveId.POISON_GAS],
-      [1, MoveId.MAGNET_RISE],
-      [4, MoveId.SMOG],
-      [7, MoveId.TAUNT],
-      [10, MoveId.ASSURANCE],
-      [13, MoveId.SLUDGE],
-      [17, MoveId.GYRO_BALL],
-      [21, MoveId.HEADBUTT],
-      [25, MoveId.SCREECH],
-      [28, MoveId.IRON_HEAD],
-      [32, MoveId.SWAGGER],
-      [36, MoveId.POISON_JAB],
-      [46, MoveId.UPROAR],
-      [52, MoveId.SPIN_OUT],
-      [58, MoveId.GUNK_SHOT],
-    ],
-    2: [
-      [EVOLVE_MOVE, MoveId.BLAZING_TORQUE],
-      [EVOLVE_MOVE, MoveId.SHIFT_GEAR],
-      [1, MoveId.LICK],
-      [1, MoveId.POISON_GAS],
-      [1, MoveId.MAGNET_RISE],
-      [4, MoveId.SMOG],
-      [7, MoveId.TAUNT],
-      [10, MoveId.ASSURANCE],
-      [13, MoveId.SLUDGE],
-      [17, MoveId.GYRO_BALL],
-      [21, MoveId.HEADBUTT],
-      [25, MoveId.SCREECH],
-      [28, MoveId.IRON_HEAD],
-      [32, MoveId.SWAGGER],
-      [36, MoveId.POISON_JAB],
-      [46, MoveId.UPROAR],
-      [52, MoveId.SPIN_OUT],
-      [58, MoveId.GUNK_SHOT],
-    ],
-    3: [
-      [EVOLVE_MOVE, MoveId.NOXIOUS_TORQUE],
-      [EVOLVE_MOVE, MoveId.SHIFT_GEAR],
-      [1, MoveId.LICK],
-      [1, MoveId.POISON_GAS],
-      [1, MoveId.MAGNET_RISE],
-      [4, MoveId.SMOG],
-      [7, MoveId.TAUNT],
-      [10, MoveId.ASSURANCE],
-      [13, MoveId.SLUDGE],
-      [17, MoveId.GYRO_BALL],
-      [21, MoveId.HEADBUTT],
-      [25, MoveId.SCREECH],
-      [28, MoveId.IRON_HEAD],
-      [32, MoveId.SWAGGER],
-      [36, MoveId.POISON_JAB],
-      [46, MoveId.UPROAR],
-      [52, MoveId.SPIN_OUT],
-      [58, MoveId.GUNK_SHOT],
-    ],
-    4: [
-      [EVOLVE_MOVE, MoveId.MAGICAL_TORQUE],
-      [EVOLVE_MOVE, MoveId.SHIFT_GEAR],
-      [1, MoveId.LICK],
-      [1, MoveId.POISON_GAS],
-      [1, MoveId.MAGNET_RISE],
-      [4, MoveId.SMOG],
-      [7, MoveId.TAUNT],
-      [10, MoveId.ASSURANCE],
-      [13, MoveId.SLUDGE],
-      [17, MoveId.GYRO_BALL],
-      [21, MoveId.HEADBUTT],
-      [25, MoveId.SCREECH],
-      [28, MoveId.IRON_HEAD],
-      [32, MoveId.SWAGGER],
-      [36, MoveId.POISON_JAB],
-      [46, MoveId.UPROAR],
-      [52, MoveId.SPIN_OUT],
-      [58, MoveId.GUNK_SHOT],
-    ],
-    5: [
-      [EVOLVE_MOVE, MoveId.COMBAT_TORQUE],
-      [EVOLVE_MOVE, MoveId.SHIFT_GEAR],
-      [1, MoveId.LICK],
-      [1, MoveId.POISON_GAS],
-      [1, MoveId.MAGNET_RISE],
-      [4, MoveId.SMOG],
-      [7, MoveId.TAUNT],
-      [10, MoveId.ASSURANCE],
-      [13, MoveId.SLUDGE],
-      [17, MoveId.GYRO_BALL],
-      [21, MoveId.HEADBUTT],
-      [25, MoveId.SCREECH],
-      [28, MoveId.IRON_HEAD],
-      [32, MoveId.SWAGGER],
-      [36, MoveId.POISON_JAB],
-      [46, MoveId.UPROAR],
-      [52, MoveId.SPIN_OUT],
-      [58, MoveId.GUNK_SHOT],
-    ],
-  },
-  [Species.PALDEA_TAUROS]: {
-    1: [
-      [1, MoveId.TACKLE],
-      [1, MoveId.TAIL_WHIP],
-      [5, MoveId.WORK_UP],
-      [10, MoveId.DOUBLE_KICK],
-      [15, MoveId.FLAME_CHARGE],
-      [20, MoveId.HEADBUTT],
-      [25, MoveId.SCARY_FACE],
-      [30, MoveId.ZEN_HEADBUTT],
-      [35, MoveId.RAGING_BULL],
-      [40, MoveId.REST],
-      [45, MoveId.SWAGGER],
-      [50, MoveId.THRASH],
-      [55, MoveId.FLARE_BLITZ],
-      [60, MoveId.CLOSE_COMBAT],
-    ],
-    2: [
-      [1, MoveId.TACKLE],
-      [1, MoveId.TAIL_WHIP],
-      [5, MoveId.WORK_UP],
-      [10, MoveId.DOUBLE_KICK],
-      [15, MoveId.AQUA_JET],
-      [20, MoveId.HEADBUTT],
-      [25, MoveId.SCARY_FACE],
-      [30, MoveId.ZEN_HEADBUTT],
-      [35, MoveId.RAGING_BULL],
-      [40, MoveId.REST],
-      [45, MoveId.SWAGGER],
-      [50, MoveId.THRASH],
-      [55, MoveId.WAVE_CRASH],
-      [60, MoveId.CLOSE_COMBAT],
-    ],
-  },
 };
